riskwise/requirements.txt
streamlit==1.40.1
groq==0.5.0
python-dotenv==1.0.0
fpdf==1.7.2
fastapi
uvicorn
